requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
jupyter
statsmodels
openai
pyyaml